4857 - InviteUserForm: use Form component
